backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.23
alembic>=1.12.1
pydantic>=2.4.2
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
requests>=2.31.0
pandas>=2.1.1
openpyxl>=3.1.2
python-dateutil>=2.8.2